fix the things this commit moves code from headers to source files, and fixes breaking with the toolchain. it also makes it easier for people to use the library by simply copying in a single directory. I hope that you find this useful :)
keep in mind that functions and variables should never be placed in header files, only prototypes. this is because of how the C preprocessor works.
