+++ .gitignore
@@ -2,7 +2,6 @@
 !.vscode/c_cpp_properties.json
 
 # Build output
-.8xp
-.bin
-.map
-main.src
+bin/
+obj/
+
